Find F given a desired level of catch

Uses the Baranov catch equation to compute the instantaneous fishing mortality rate (F) that would yield a specific level of catch_

find_F(f_guess, naa, waa, mort, selex, ret = NA, dmr = NA, prov_catch)

Arguments

f_guess

initial guess for F

naa

numbers-at-age

waa

weight-at-age

mort

natural-mortality-at-age

selex

selectivity-at-age

ret

retention-at-age (optional)

dmr

instantaneous discard mortality rate (optional)

prov_catch

desired catch level

Value

the fishing mortality rate that yields the desired catch level

Examples

if (FALSE) { # \dontrun{
find_F(f_guess=0.1, naa=naa, waa=waa, mort=mort, selex=selex, prov_catch=0.12)
} # }